requirements.txt
Package            Version
------------------ -----------
brotlipy           0.7.0
certifi            2022.5.18.1
cffi               1.15.0
charset-normalizer 2.0.4
cryptography       37.0.1
cycler             0.11.0
fonttools          4.33.3
idna               3.3
joblib             1.1.0
kiwisolver         1.4.3
matplotlib         3.5.2
mkl-fft            1.3.1
mkl-random         1.2.2
mkl-service        2.4.0
numpy              1.21.6
packaging          21.3
pandas             1.1.5
Pillow             9.1.1
pip                21.2.4
protobuf           3.20.1
pycparser          2.21
pyDeprecate        0.3.2
pyOpenSSL          22.0.0
pyparsing          3.0.9
PySocks            1.7.1
python-dateutil    2.8.2
pytz               2022.1
PyYAML             6.0
requests           2.27.1
scikit-learn       1.0.2
scipy              1.7.3
setuptools         61.2.0
six                1.16.0
sklearn            0.0
tensorboardX       2.5
threadpoolctl      3.1.0
torch              1.10.2
torchaudio         0.10.2
torchmetrics       0.8.1
torchvision        0.11.3
typing_extensions  4.2.0
urllib3            1.26.9
wheel              0.37.1
win-inet-pton      1.1.0
wincertstore       0.2
